auth_React/fast_api/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
pydantic-settings
passlib
python-jose
aiofiles
python-multipart
celery[redis]
redis

